Reorganise efficiency, and spin out export functions
